requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
httpx==0.26.0
PyJWT==2.8.0
cryptography==41.0.7
itsdangerous==2.1.2
aiofiles==23.2.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
APScheduler==3.10.4
pymysql==1.1.0